Add earthquake heatmap example
This uses the lastest from ios-v3.4.0-alpha.4 to cluster geojson
source data and apply styles to filtered sets of that data. It also
updates all examples to not animate the map when it appears -- this
is required since the activities are loaded at viewDidLoad which is
too early to set a map's center with an animation.